refactor: 스케줄러 수정
